requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
jinja2==3.1.4
python-docx==1.1.2
reportlab==4.2.5
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
anthropic>=0.40.0
httpx>=0.27.0
openai>=1.50.0
google-generativeai>=0.8.0
pdfplumber>=0.11.0
psycopg2-binary>=2.9.9
